+++ suitkaise/sk/__init__.py
@@ -28,5 +28,5 @@ __all__ = [
 ]
 
 # Module metadata
-__version__ = "0.4.4b0"
+__version__ = "0.4.5b0"
 __author__ = "Suitkaise Development Team"